requirements.txt
jupyterlab
python-dotenv
langchain
openai
langchain-openai
langchain_core
langchain_community
langchain_huggingface
langchain_text_splitters
faiss-cpu
python-dotenv
torch
openai
fastapi
uvicorn
pydantic
motor
selenium
bs4
passlib[bcrypt]
pyjwt
langchain_teddynote